Address various issues when stdout is not a tty
Do not (try to) prompt for password in the above situation. Raise an
appropriate warning/error, and create the '-p' option for command-line
passwording instead, while passwording through the config file remains
an option. Ensure the former takes precedence, even when the password
is empty, which requires a workaround.
Colorize password prompts and errors suitably.